+++ convergence_analyses/render_figure_energy_decay_space_refinement.py
