+++ convergence_analyses/render_figure_energy_decay_theta.py
